Refactor tank moving and collision handling methods
The tank moving and collision handling methods have been refactored to improve readability and maintenance. Previously, the 'move' method in CTank class was cluttered and was responsible for multiple tasks i.e., moving the tank, checking for collisions, and handling them.

Now, the logic has been subdivided into smaller methods ('moveInDirection', 'handleObjectCollision', 'restorePosition') each with their distinct responsibility, adhering to the Single Responsibility Principle. This minor change improves the clarity of the code making it easier to interpret and manage. There is no change in the functionality of the program.

The order of import statements in CBullet was also corrected, and the Q_DECLARE_INTERFACE macro was added to the IBase class, ensuring compatibility with the Qt meta-object system.